Time: 25 ms (43.65%), Space: 45.1 MB (60.18%) - LeetHub

+++ 0875-koko-eating-bananas/0875-koko-eating-bananas.java
@@ -10,7 +10,7 @@ class Solution {
     private int calculateTotalHours(int[] piles,int hourly){
         int  totalHours=0;
         for(int pile: piles){
-            totalHours+=(int) Math.ceil((double)pile/(double)hourly);
+            totalHours+= Math.ceil((double)pile/(double)hourly);
         }
      return totalHours;   
     }
